complete deco project, refactor naming for consistency

+++ utils/transform.py
@@ -3,7 +3,9 @@ from utils.matplotlibutils import path_from_coordinate_list, coordinate_list_fro
 import itertools
 
 
-def rotate(path: [(float, float)], degrees: int, rotation_x: float, rotation_y: float):
+def rotate(
+    path: [(float, float)], degrees: float, rotation_x: float, rotation_y: float
+):
     mpl_path = path_from_coordinate_list(coordinate_list=path)
     rotation = transforms.Affine2D().rotate_deg_around(rotation_x, rotation_y, degrees)
     mpl_rotated_path = mpl_path.transformed(rotation)
